+++ src/__tests__/standalone-test.js
@@ -304,7 +304,7 @@ test("standalone extending a config that is overridden", t => {
 test("standalone loading YAML with custom message", t => {
   standalone({
     code: "a { color: pink; }",
-    configFile: path.join(__dirname, "fixtures/config-color-no-named-custom-message.yaml"),
+    configFile: path.join(__dirname, "fixtures/config-color-named-custom-message.yaml"),
   }).then(({ output }) => {
     const parsedOutput = JSON.parse(output)[0]
     t.equal(parsedOutput.warnings.length, 1)