+++ src/postcssPlugin.js
@@ -72,22 +72,8 @@ export default postcss.plugin("stylelint", (options = {}) => {
         // Ignore the rule
         if (primaryOption === null) { return }
 
-        let ruleSeverity
-        if (get(secondaryOptions, "warn")) {
-          result.warn((
-            "'warn: true' has been deprecated and in 5.0 it will be removed. " +
-            "Use 'severity: \"warning\"' instead."
-          ), {
-            stylelintType: "deprecation",
-            stylelintReference: "http://stylelint.io/user-guide/configuration/",
-          })
-          ruleSeverity = "warning"
-        } else {
-          ruleSeverity = get(secondaryOptions, "severity", "error")
-        }
-
         // Log the rule's severity in the PostCSS result
-        result.stylelint.ruleSeverities[ruleName] = ruleSeverity
+        result.stylelint.ruleSeverities[ruleName] = get(secondaryOptions, "severity", "error")
         result.stylelint.customMessages[ruleName] = secondaryOptions && secondaryOptions.message
 
         // Run the rule with the primary and secondary options